requirements.txt
tensorflow>=2.12
streamlit
pillow
scikit-learn
numpy
matplotlib
seaborn